Make Q&A questions list scrollable
