+++ src/components/q-and-a/constants.ts
@@ -1,0 +1,119 @@
+interface QAndAQuestion {
+  question: string;
+  hint: string;
+  answer: string;
+  description: string;
+}
+
+export const qAndAContent: QAndAQuestion[] = [
+  {
+    question: "Where does Rainforest Foundation US work?",
+    hint: "I know!",
+    answer:
+      "RFUS works together with Indigenous peoples in South America’s Amazon Rainforest.",
+    description:
+      "Specifically, our programs are in Peru, Brazil, Guyana, and Panama. We also have region-wide initiatives across Mesoamerica, the entire Amazon rainforest, and the whole tropical belt of the planet!",
+  },
+  {
+    question:
+      "What does Rainforest Foundation US (RFUS) do? How does RFUS stop people from cutting down trees in the rainforest?",
+    hint: "Let me explain...",
+    answer:
+      "RFUS supports Indigenous communities in the Amazon Rainforest by in five key ways:",
+    description: `
+<b>1. Territorial Monitoring:</b>
+RFUS trains community members to use cellphone apps, GPS technology, and drones to detect deforestation in real time, report it to authorities, and kick out the people responsible!
+
+<b>2. Land Management:</b>
+RFUS works with communities to build systems that protect natural resources and local cultures, and also offer people sustainable and dignified employment opportunities. 
+
+<b>3. Policy & Advocacy:</b>
+RFUS acts like a bridge, connecting local communities in the Amazon and decision-makers around the world, so that laws are written to empower Indigenous peoples, and protect the natural resources that  we all depend on (like the trees of the forest)! 
+
+<b>4. Institutional Strengthening:</b>
+RFUS invests directly in Indigenous-led organizations so they can protect their forest and support their own people in the ways they know best. 
+
+Indigenous peoples understand the problems facing the Amazon Basin more than anyone else. We, the people of the planet, need them to lead us in finding solutions! Placing money directly in Indigenous hands helps make this possible.
+
+<b>5. Land Titling and Legal Intervention:</b>
+RFUS helps pay legal fees for Indigenous peoples and Indigenous organizations, whenever this is necessary. Court cases are expensive!`.trim(),
+  },
+  {
+    question: "Why does RFUS want to work with Indigenous peoples?",
+    hint: "That’s a great question!",
+    answer:
+      "RFUS works with Indigenous peoples and local communities because they are the world’s best forest defenders!",
+    description:
+      "Being on the frontline of deforestation, they are most impacted by it, and they know best how to stop it.",
+  },
+  {
+    question:
+      "How come outsiders want to cut down the rainforest, and Indigenous people don’t?",
+    hint: "It’s complicated…",
+    answer:
+      "There are many reasons people and organizations (like companies, and even governments) want to clear trees in the rainforest.",
+    description: `
+Sometimes, they are clearing space to build mines, and extract valuable resources (like the precious metals inside e computers and phones); other times they create large-scale agricultural projects that produce food—often meat—for people around the world . 
+
+Other times, people go to the Amazon because they feel they have nowhere else to go,  and start small farms just to feed themselves and their families. 
+
+It’s important to note that not all Indigenous people grow up and become forest defenders. Some pursue totally different career paths, and a few even decide to help cut down forests themselves.Overall, though, Indigenous peoples have the knowledge governments around the world need to make rules and build systems that best protect forests.`.trim(),
+  },
+  {
+    question:
+      "If Indigenous peoples were in the Amazon first, why can outsiders cut down the forest? Why don't the police do anything to help?",
+    hint: "This is a hard one…",
+    answer:
+      "The Amazon is very large. As such, it is difficult for governments, institutions, and local communities to enforce laws; and easy for people and organizations to show up in the Amazon, claim land as their own, and deforest.",
+    description:
+      "This is why it is essential for Indigenous peoples to have titles to their lands. With land titles, they can prove to authorities that the land is theirs, and that others cannot come to the forest, cut down trees, build homes and farms, and claim ancestral Indigenous territory for themselves.",
+  },
+  {
+    question:
+      "What will happen to the Indigenous people if the rainforest gets cut down?",
+    hint: "I’m really glad you asked.",
+    answer:
+      "Indigenous peoples will lose the home they have been living in for thousands and thousands of years.",
+    description:
+      "They will lose their culture and language, too. This would be tragic.",
+  },
+  {
+    question:
+      "What will happen to people around the world if the rainforest gets cut down?",
+    hint: "A very important question!",
+    answer:
+      "If rainforests around the world get cut down, the world loses the largest land-based carbon sink.",
+    description:
+      "The rainforest would reach a tipping point, and become a grassland. Countless plants and animals would disappear from the face of the earth, and the impact of climate change - from drought to extreme storms - would increase, all around the world. This would be devastating.",
+  },
+  {
+    question: "What happens when I donate to RFUS?",
+    hint: "Awesome question!",
+    answer:
+      "Your donation funds all of the work RFUS does, from organizing fundraisers to supporting Indigenous peoples to get titles to their ancestral land.",
+    description:
+      "Explore our website to get a better sense of what we do, and how your donation makes a difference!",
+  },
+  {
+    question:
+      "What can I do to help the rainforest, and the communities living there?",
+    hint: "I’m so glad you want to know! Here’s some ideas:",
+    answer: "There’s plenty you can do to protect the rainforest!",
+    description: `
+The good news is, you’re already taking the first step: educating yourself about the global importance of the Amazon Rainforest, and about the role Indigenous peoples play in protecting it.
+
+Once you understand deforestation, climate change, and the challenges facing Indigenous peoples, think about why these problems exist. Making changes in your own behavior is a great way to start, but when you work with others, you can have a massively bigger impact. And it’s important to remember that some people have more ability, power, or flexibility to make change than others. 
+
+Here’s some things you can think about on your own: 
+
+Certain ingredients require deforestation to be produced. Common examples are palm oil, beef, and soybeans (for soy milk, but also for chicken feed). Do your best to consume these products less. 
+When we burn fossil fuels like oil and gas, we send harmful gasses into the air, and we contribute to climate change. This harms the Amazon. Instead of traveling by car, can you walk, or bike? Instead of flying far away for your next vacation, can you find something fun to do that is local? 
+
+Here are a few actions you can take with others: 
+Share what you are learning with friends and family, and help them get involved in making a positive difference.
+Get political! Call your local newspaper, and ask them to publish more news about what is going on in the Amazon. Call your local representatives, and let them know that protecting the Amazon is important to you, and that it should be a top priority for them, as well!
+Start a fundraiser at your school or your local community center! If you choose to do this, reach out to RFUS, and we will give you the tools you need to be successful!  
+
+You’re on an exciting path - keep learning, and striving to make a difference!`.trim(),
+  },
+];
